added new predefined Real class for floating point calculations
git-svn-id: svn+ssh://agate.cs.unh.edu/home/csg/nau7/svn/mate@24 8533781a-2b67-4406-8f2e-75780de0317d

+++ src/mvm/thread.h
@@ -1,5 +1,5 @@
 /* Niels Widger
- * Time-stamp: <19 Dec 2010 at 18:32:11 by nwidger on macros.local>
+ * Time-stamp: <20 Dec 2010 at 12:26:11 by nwidger on macros.local>
  */
 
 #ifndef _MVM_THREAD_H
@@ -51,13 +51,13 @@ uint32_t _thread_get_pc(struct thread *t);
 
 uint32_t thread_set_pc(uint32_t p);
 
-uint32_t thread_get_instruction_counter();
+uint64_t thread_get_instruction_counter();
 
-uint32_t _thread_get_instruction_counter(struct thread *t);
+uint64_t _thread_get_instruction_counter(struct thread *t);
 
-uint32_t thread_increment_instruction_counter();
+uint64_t thread_increment_instruction_counter();
 
-uint32_t _thread_increment_instruction_counter(struct thread *t);
+uint64_t _thread_increment_instruction_counter(struct thread *t);
 
 int thread_start(struct object *o);
